+++ NEWS.md
@@ -1,3 +1,16 @@
+# ghqc.app 0.6.0
+
+`ghqc_status_app()`:
+
+- table buttons are now in an action drop-down menu
+- action buttons exist to perform the actions:
+    
+    - Approve
+    - Notify file changes
+    - Notify latest commit
+    - Repost last QC notification
+    - Unapprove
+
 # ghqc.app 0.5.3
 
 - fixes bug in `ghqc_record_app()` in which in warning modal still notifies when there are no issues with unchecked checklist items